Update PAA5100JE.py
Final version after testing: both sensors were successfully initialized and were able to read motion, however the read data were not send to the pyboard. A possible cause would be an error during the ISR, because no ID were provided by this sensor to allow it to perform an interrupt.